requirements.txt
meshtastic
flet
pyserial